move get_default_consumption to shortcuts
