+++ corehq/apps/consumption/models.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 from couchdbkit.ext.django.schema import Document, StringProperty, DecimalProperty
 
 
@@ -17,18 +16,3 @@ class DefaultConsumption(Document):
     supply_point_type = StringProperty()
     supply_point_id = StringProperty()
     default_consumption = DecimalProperty()
-
-
-def get_default_consumption(domain, product_id, location_type, case_id):
-    keys = [
-        [domain, product_id, {}, case_id],
-        [domain, product_id, location_type, None],
-        [domain, product_id, None, None],
-        [domain, None, None, None],
-    ]
-    results = DefaultConsumption.get_db().view(
-        'consumption/consumption_index',
-        keys=keys, reduce=False, limit=1, descending=True,
-    )
-    results = results.one()
-    return Decimal(results['value']) if results else None
